django/requirements.txt
Django
gunicorn
djangorestframework
pillow
django-cors-headers
PyYAML
psycopg2-binary
rpdb